Repackage firebase-web module.
