+++ firebase-web/src/main/java/io/spine/web/firebase/query/FirebaseQueryProcessingResult.java
@@ -18,8 +18,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.web.firebase;
+package io.spine.web.firebase.query;
 
+import io.spine.web.firebase.FirebaseDatabasePath;
 import io.spine.web.query.QueryProcessingResult;
 
 import javax.servlet.ServletResponse;
