+++ firebase-web/src/main/java/io/spine/web/firebase/query/FirebaseQueryServlet.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.web.firebase;
+package io.spine.web.firebase.query;
 
 import io.spine.web.query.QueryServlet;
