+++ firebase-web/src/test/java/io/spine/web/firebase/query/FirebaseQueryBridgeTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.web.firebase;
+package io.spine.web.firebase.query;
 
 import com.google.api.core.ApiFuture;
 import com.google.firebase.database.DatabaseReference;
@@ -30,8 +30,8 @@ import io.spine.base.Time;
 import io.spine.client.Query;
 import io.spine.client.QueryFactory;
 import io.spine.testing.client.TestActorRequestFactory;
+import io.spine.web.firebase.query.given.FirebaseQueryMediatorTestEnv.TestQueryService;
 import io.spine.web.query.QueryProcessingResult;
-import io.spine.web.firebase.given.FirebaseQueryMediatorTestEnv.TestQueryService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -40,9 +40,9 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeoutException;
 
 import static io.spine.json.Json.toCompactJson;
-import static io.spine.web.firebase.given.FirebaseQueryBridgeTestEnv.nonTransactionalQuery;
-import static io.spine.web.firebase.given.FirebaseQueryBridgeTestEnv.transactionalQuery;
-import static io.spine.web.firebase.given.FirebaseQueryMediatorTestEnv.timeoutFuture;
+import static io.spine.web.firebase.query.given.FirebaseQueryBridgeTestEnv.nonTransactionalQuery;
+import static io.spine.web.firebase.query.given.FirebaseQueryBridgeTestEnv.transactionalQuery;
+import static io.spine.web.firebase.query.given.FirebaseQueryMediatorTestEnv.timeoutFuture;
 import static java.util.Collections.singletonList;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.any;
@@ -62,7 +62,8 @@ import static org.mockito.Mockito.when;
 class FirebaseQueryBridgeTest {
 
     private static final QueryFactory queryFactory =
-            TestActorRequestFactory.newInstance(FirebaseQueryBridgeTest.class).query();
+            TestActorRequestFactory.newInstance(FirebaseQueryBridgeTest.class)
+                                   .query();
     private static final int ONE_SECOND = 1000 /* ms */;
     private static final int SECONDS = ONE_SECOND;
 
@@ -88,7 +89,8 @@ class FirebaseQueryBridgeTest {
                                                               .setDatabase(firebaseDatabase)
                                                               .build();
         final Query query = queryFactory.all(Empty.class);
-        final QueryProcessingResult result = bridge.send(nonTransactionalQuery(query));
+        final QueryProcessingResult result = bridge.send(
+                nonTransactionalQuery(query));
 
         assertThat(result, instanceOf(FirebaseQueryProcessingResult.class));
     }
